Replace native git use with JGit outside of tests
We also upgrade to Gradle 8.5, allowing us to use test fixtures.

We still use the `git` executable for setting things up, but the
configuration cache doesn't like us running external executables.  JGit
doesn't support quite the right describe features out of the box, so we
reimplement what we need for consistency with the previous version.

Worktree support is dropped.

Signed-off-by: Andrew Aylett <andrew@aylett.co.uk>