Use Gradle's provider factory executor to run Git
This means we may keep better track of what _exactly_ we're using to
key the configuration cache.  Egit had some situations where it needed
to shell out to system git, and Gradle didn't like that.  But calling
system git through Gradle APIs is fine, and avoids Gradle needing to
use the content of Git's repo in its cache key.

Signed-off-by: Andrew Aylett <andrew@aylett.co.uk>